Remove dummy test to boost projection performance.
Mpop was checking in 2 different places if the source and target areas were
different, leading to pyresample expanding the area definitions to full
lon/lat arrays when checking against a swath definition, and then running
an allclose. This was inefficient, and the programming team decided that it
was the user's task to know before projection if the source and target area
were the same. In other words, the user should be at least a little smart.

Signed-off-by: Martin Raspaud <martin.raspaud@smhi.se>