Move composite loading out of Scene to mpop.composites
The loading of compositors was a part of the Scene object. However, it does
not belong there, so we decided to move it out of Scene. The next logical
place to have it is the mpop.composites modules.
As a conterpart, we now provide the `available_composites` method to the
Scene to be able to figure out what we have possibility to generate.

Signed-off-by: Martin Raspaud <martin.raspaud@smhi.se>